Implemented a system to read from an xml file.
The program should be able to sucessfully read an xml file, and input
data from the course there based on the selection of the course at
menu2. Currently we have only 1 course thats inputted into the xml
file, but we can add many more if needed. The program does the xml
reading in the background when the user hits the Start game button on
menu 4.