Interface changes made to the Portrait screen
Rearranged the look and feel of the UI on the Portrait screen. In
Addition, a table was implemented to display the player names inputed
from menu 1.

Most of the UI changes are experimental, and will require revisions.
Need to figure out the best way to use screen real estate and have it
be easy for the user to interact with.