Portrait mode has the functionality to input scores and have them show on screen
Many additions were made to the Portrait mode today.

- Changed the name of the Bet column label to Score, as this shows what
the player score is for that hole.

- In menu 2, Bethpage is selected by default, so it doesnt crash
anymore. You should select bethpage every time for it is the only
course set up properly.

- When you press next hole, it changes the hole number, par and
handicap for the next hole appropiately. The previous hole button works
as well. This may be temporary as finger swipes could replace these.

- You can now input the scores for each player on a hole. It is just
general functionality that updates the screen, and doesnt change with
which hole we are on yet. This will probably be the next piece of
functionality to be added.