Added side betting point value inputs plus small update on the transition to ios7.
This is the first verbose push since xamarin was updated to  support
ios 7. Just to point a few things out:
-The UI has changed alot in look and feel. Buttons have changed,
swtiches are more simple, etc.
-Screen size has changed as well. The simulator used to test the app is
the iphone 4 inch screen which is a different aspect ratio from
previous models.

Menu 4 now contains a field for each side bet for the user to enter.
The options are only shown when side bets are turned on. If the side
bets swtich is off, the inputs go invisible.

A go back button was added to Portrait mode so that the user can go
back to the setup phase.