Further devlopment on the landscape scorecard
There is now a scrollview with collectionViews laid out as a grid. The
values are just static values currently. Nothing is curently linked up.
Some alignment is required for both segments of the scroll view.

This is not entirely the intended design for this screen, so design
changes may be need for overall design.