Significant progress in the betting system.
It would appear that betting with a 4 player team, seems to function.
This needs to be further tested. The logic is there for 4 players but
not for any other scenarios. It is also possible that the logic doesnt
satisfy all situations of a bet or stages the app is in.