More development to the betting tab
The font for the betting tab UIlabels has been changed to
Helvectica-Bold with a point of 16. This was done so that double digit
values could be properly displayed in the grid cell. We can change the
font to anything we want in the future.

Also the cells will include a + or - symbol next to the value as
indicated on the google spreadsheet.

The betting functionality a fairly functional base. You can set point
values to multiple bets, input them in portrait mode, and then see them
properly added up in landscape mode. Frank should examine and test the
betting part to make sure it matches our original design plan.