Handicap color values added
The handicap colors have now been implemented in. They should only show
up next to the score when a handicap applies to that hole.

This implementation is not fully tested, as there could be a number of
bugs. There is also some instances that need to be resolved.

For Example: A hole has a par of 3. A player has a handi of 18 and this
hole has a handicap of 12. The player gets a hole in one. Does this
mean he gets an extra stroke which reduces his stroke count to 0?