Added a Segmented control to menu 1 for number of players.
A uisegmentedcontrol was added to menu 1 to allow the user to choose
the number of players. Valid choices are 1 to 4 players. This feature
is currently untested. So for stable results its best to leave it at 4
players.

The control is setup to hide/show the player name slots. This prevents
input going in the wrong text fields.