Totals for the nassau added on the betting tab.
The second total for the lower nine on the nassau game is now partially
working. The first total will keep track of side bets while the second
total combines all the holes a player has won. This is only for the
bottom nine holes at this point in time.

To my knowledge, nassau is played where each hole is a point. If theres
a tie for the lowest stroke, no one wins that hole. Presses are being
ignored right now as I am trying to keep this simple.