Implemented first stages of joint betting scores on betting tab for nassau
The nassau game now has a joint score mode in the betting tab. Showing
both your side bet score on the left, and the nassau point(should only
be a 1 or a 0) on the right. The text is probably too small to see.
Will need to make this bigger later.