Added Total labels for nassau and modified the grid for par in skins.
-Added a Total label for each player in nassau mode. The total simply
adds the out and in for the nassau bets.

-Added some features to the skins mode. The par grid has been mutated
to show the current skin value for that hole and the pushes if any
exist. For example if theres a skin of $2 on a hole it will show a
value of "2". However if theres a value of 4 thats been pushed to that
hole it will show "2+4".

These changes are somewhat done indepentantly of my own decision. I
felt we need to show the skin value somewhere, so I simply replaced the
par row with skin value.

These changes/additions need to be reviewed to make sure its how we
want it designed.