Wolf updates added.
Some implementations to wolf was added. Right now its designed so that
after each score input, it asks the player if it wants to become part
of the wolf pack for that hole. However, this is incorrect, as a player
becomes part of a wolf pack on first drive not after the final score is
inputted.

So this will likely change in the next build to something of a toggle
on the portrait view next to the player's name.