Updated the wolf interface for portrait view.
The Add WP view has been removed. There is now WP buttons for each
player that is not the wolf. They act like toggles and use the same
images for the score buttons. (As always, UI can be stylized down the
road later.)

Currently the WP toggles do not stick when the user selects the next
hole. This may or may not change depending on design decisions. This
new UI added should allow us to properly play a wolf game before
entering scores.